Switch JSON schema validator
The current one depends on the non-free org.json (see
https://wiki.debian.org/qa.debian.org/jsonevil ), which also has an
unfortunately clash with the alternative re-implementation of org.json
used in spring (although my attempt to workaround the latter seems to
have been successful).

The networknt one was picked from:

https://json-schema.org/implementations.html

on the basis that it seems to be the next most popular java
implementation based on github stars/forks.

It's not clear why we had exclusions for jackson-databind (which the
new schema validator depends on).

They were originally added by Justin in commit
3c932929ac2e65e265f3393ef9006c0b8b6d419f in September 2017, but no
reason was given.

Even once un-excluded, there was still an issue with version, I
believe as maven was probably picking to use the version spring
requested which was too old, resulting in an error when trying to run
the unit tests:

java.lang.NoSuchMethodError:
com.fasterxml.jackson.core.JsonFactory.requiresPropertyOrdering()Z

So an explicit dependency on a suitable version of jackson was added.

The choice of version 2.12.1 is because that's what the validator
currently seems to list in it's dependencies:

https://github.com/networknt/json-schema-validator/blob/master/pom.xml

The implementation is also split up a little to make it easier to add
verification of requests too in the future.