ekyc: Change tests to derive from AbstractOIDCCServer
Currently it derives from OIDCCServer, which includes extra checks -
including the one that cause #988.

This removes some other checks as well - however these are all checks
on the underlying OIDCC protocol (like verifying at_hash or
authorization code lengths). This seems correct; the goal of the
ekyc tests is to verify the ekyc related part of the protocol and it
is recommended/expected that a security profile certification (e.g. a
FAPI one) is run too.

closes #988