Run payment tests against mockbank in CI
Relies on a API access token manually created and added to the GitLab
CI settings (this particular one was created by logging into one of my
gitlab accounts that's not an admin user for the suite).

One test fails because it's just not possible to pass it on the
mockbank; other tests have 'override' sections adding in the config
such that it uses magic payment amounts as per:

https://gitlab.com/obb1/certification/-/wikis/Phase-3-Payments

to trigger the correct response from the mock bank.

I didn't know how to wait for the newly pushed version of the suite to
deploy, so a 'sleep 120' was used.

gitlab-ci.yml's 'resource_group' syntax is used to ensure no other
deployment (or test) jobs run inbetween deployment and the tests
starting, or whilst the tests are running. For this reason the tests
are made part of the upload job (if they were two separate jobs, a
deployment job from a different pipeline might run after our
deployment but before our tests). [This is because we only have a
single 'dev' environment that all branches deploy to, rather than
having one unique review app for each branch.]

Fixes George had made to
brazil-raidiam-functional-payments-automated.json in a different
branch are incorporated.