Avoid leaking htmlunit/selenium webclients
I noticed that after running some tests locally and pausing the JVM,
inspecting the thread list in IntelliJ revealed a number of htmlunit's
JS executor threads running.

Googling led to pages like this:

https://stackoverflow.com/a/64183368/292166

where it became clear it's important to call the 'close()' method
on webclient - which in htmlunit seems to be accomplished by calling
close on the driver.

This seems to stop the accumulation of javascript executor threads
(and presumably other things).

This is a possible cause of the OOM that caused outage #1016