Avoid interacting with selenium from multiple threads
AbstractTestModule was calling getWebRunners() to try and figure out
if any runners are active. Unfortunately this calls several things
that poke down into Selenium, e.g. to get the current url - and now
that we close selenium down, this can happen whilst it is being closed
down.

Solve this by having a 'are there runners active' function that
doesn't need to call into Selenium.

The web frontend potentially stills calls into here, so there may
still be chances for that to see errors. The code doesn't look thread
safe, in that we may end up querying a driver at the same time another
thread decides to start destroying it. It's not clear if this might
cause issues or not - we have been accessing the htmlunit/driver in
this way from multiple threads since it was first added, and the
documentation makes clear that it's not thread safe. ("Note: a
WebClient instance is not thread safe. It is intended to be used from
a single thread." and so on.)