Remove dependency of Brazil client rs conditions on test module
Refactor so that the resource server conditions do not depend on the
test module.

We want to share these conditions across the FAPI1 and FAPI2 test
modules, so having a dependency on the test module is asking for
trouble when things potentially get changed in the future.

It probably would have been neater to put the required path into the
environment instead.