Create initial FAPI2Baseline tests, as exact copies of FAPI1Adv
Achieved by using IntelliJ to rename the
net.openid.conformance.fapi1advancedfinal package to fapi2baselineid2,
then undoing the deletions - on the basis these tests are the closest
thing we have to fapi2, so are a good starting point.

This follows the method used to create the FAPI1Adv Final tests in
https://gitlab.com/openid/conformance-suite/-/merge_requests/987

We are going to target implementers draft 2 of the fapi2baseline
standard - which doesn’t actually exist yet, so we’re technically
working to the latest draft on master in the fapi specs git repo.

Further commits will do more renaming (of the filenames/class
names/test names/etc) and start to update the tests to actually test
the FAPI2 standard.

Part of #1017