Workaround timeouts when calling exporthtml endpoint
The CI was occasionally ending up with broken zip files, particularly
for the ekyc job which runs when there are lots of parallel tests
going on.

Checking the GCP console revealed this exception occurring:

org.springframework.web.context.request.async.AsyncRequestTimeoutException

Somewhat frustratingly, when this exception occurs the download is
cleanly truncated instead of being aborted (at the HTTP level, the
java docker container is using the chunked transfer encoding, and
cleanly finishes by sending the whole of a chunk, then sending the
expected zero length chunk of a successfully completed download).

This increases the timeout for all APIs; increasing it just for the
problem exporthtml API would have been preferable but spring doesn't
appear to have an easy mechanism for doing so.

See:

https://docs.spring.io/spring-boot/docs/current/reference/html/application-properties.html#application-properties.web.spring.mvc.async.request-timeout

The 2 minutes was picked on the basis I saw the ekyc export job take
24 seconds in the CI, and then allowing a fair amount of leeway above
that.