run-test-plan.py: Detect if we download a corrupt zip file
The CI was occasionally ending up with broken zip files, particularly
for the ekyc job which runs when there are lots of parallel tests
going on.

Checking the GCP console revealed this exception occurring:

org.springframework.web.context.request.async.AsyncRequestTimeoutException

Somewhat frustratingly, when this exception occurs the download is
cleanly truncated instead of being aborted (at the HTTP level, the
java docker container is using the chunked transfer encoding, and
cleanly finishes by sending the whole of a chunk, then sending the
expected zero length chunk of a successfully completed download).

To avoid us ending up with the corrupt zip files being used in the CI
compare job, detect them and retry the download. (The actual
underlying problem of the timeout is being fixed in another commit by
increasing the timeout.)