Reduce parallelism in server_test job at start
We end up running quite a lot of tests at the start of the server_test
job as we parallelise all those without an alias. This isn't the
slowest queue so this likely just overwhelms the server a bit and
slows down other tests (hence probably why we saw exporthtml timing
out, particularly in the ekyc test jobs - server/database load was
excessive).

Run the jobs without an alias sequentially to avoid this. We leave
local_test running the jobs in parallel though, as all it's jobs have
no alias and it's running on it's own instance of the suite.