FAPI2Baseline: Change tests to use response_type=code + remove JARM
This brings the tests almost into alignment with the specification
I believe (the major remaining difference being there should at least
be an option to use unsigned request objects).

Various new conditions are added to make sure the response is in
the right place (i.e. in the url query and not, as in FAPI1, in the
url fragment/hash).

The existing FAPIRWCheckDiscEndpointJARMResponseTypesSupported is
renamed to a more generic name, as all it does is check the 'code'
response type is supported.

part of #1017