Disable /api/info endpoint
This endpoint returns all of the test modules a user has run
(or for an admin, all test modules in the system).

Unfortunately the returned list quickly gets very large, and
there's no pagination mechanism. A user of the suite appears to
be regularly calling this API, and there test list has grown
large enough that retrieving it now causes out of memory exceptions
in the test suite as it pulls many tens of thousands of TestInfo
entries into RAM.

This endpoint isn't used in our own web ui, and I'm not convinced
there is a real use case for using it. (If there is, we could apply
the pagination method applied to other APIs and re-enable it.)