Update nimbus to the latest version
This just seemed a sensible thing to do as we're quite behind and the
newer version has some potentially useful helper methods etc (as well
as some potential security fixes).

This is a major version upgrade so there are some API changes; in
particular the minidev json smart library is no longer part of the
nimbus public API, so remove our uses of it.

Upgrade tink to the version nimbus say they depend on (it's an
optional dependency so we have to explicitly list it).