Refactor: Remove CallProtectedResourceWithBearerTokenAndCustomHeaders
This only very slightly differs from
CallProtectedResourceWithBearerToken - it just allows customs headers
to be overridden. This could also be achieved by just not setting
custom headers if none are asked for, so take that approach.

It also align it's behaviour with
CallProtectedResourceWithBearerTokenExpectingError, which optionally
supports custom headers.

To verify this wouldn't change any behaviour, I first ran the full CI
with CallProtectedResourceWithBearerToken set to throw an error if
resource_endpoint_request_headers was present in the environment, and
nothing threw - indicating we were ever used it to deliberately not
sent custom headers that were set in the environment.

The aim here is to reduce the number of similar CallProtectedResource
conditions.