Add a new "CallProtectedResource"
This has different behaviour CallProtectedResourceWithBearerToken,
namely that it explicitly treats all http status codes as success,
so can be used regardless of whether failure or success is expected.

This will make the error handling more explicit and mean we can
remove CallProtectedResourceWithBearerTokenExpectingError in
due course.

The name is picked on the basis this will shortly also support
dpop tokens, so a name without 'bearer' makes sense - and I wanted
not to change the behaviour of the existing condition as the
folks that have forked our codebase to create functional tests may
be relying on the current behaviour.