Remove some use of CallProtectedResourceWithBearerTokenExpectingError
We've slowly been migrating the code to a model where the conditions
calling endpoints over HTTP just do the call, returning the full
response and then further condition calls are used to verify the http
status etc are as expected.

Apply that model here by using CallProtectedResource when checking if
the access token was revoked when reusing an authorization code.

There are slight differents in the validation being done:

we now explicitly check the http status is the expected 401 error
(whereas previously I believe any 4xx or 5xx code was accepted).

the old condition allowed for a 200 response containing a json object
with an error field; the logic for allowing this is unclear (I cannot
find any standard that mentions it and the code has been there since
the early days of the OBIE UK tests) - the new code does not allow
this option (though it could easily be added if we find a reason for
doing so).

We have multiple shoulds in play here meaning we only use a 'warning'
- revoking the access token is a 'should', and returning 401 for a
revoked access token is a 'should'.

We could perhaps check for the presence of the WWW-Authenticate http
header (which
https://datatracker.ietf.org/doc/html/rfc6750#section-3.1 says is a
'MUST') but the DPoP language seems a bit vaguer (see
https://github.com/danielfett/draft-dpop/issues/119 ) so I'm not sure
we can apply that in all cases. For now I've left the behaviour as per
the existing code and WWW-Authenticate is not checked.

Equally we go for any 4xx error as showing the token was revoked -
this is slightly stricter than the existing code as it removes the
possibly of a 500 response.