Remove everywhere we explicit request utf8 charset for json
These are almost all on the basis of a widespread misunderstanding
that had perviated into a multitude of places, including Spring, early
draft of fapi specs and more.

The default charset for the json mimetype is utf8, so there is no need
to explicitly request it.

We do have some tests that are deliberately trying different variants
of the Accept header, and these are now trying the 'plain' (without
charset) variant twice - my next commit will fix those to try testing
the utf8 variant, as it is still valid (albeit verbose) and servers
should continue to be tested to accept it.