For tests that support Brazil payments, allow 201 from resource endpoint
We might be creating a payment with POST hence a 201 is allowed,
but in other cases it'll be a 200.

Making this an explicit test that it's 201 only in the case where
we did a POST would probably be better, but for now this is better
than what we had before.