Clear token endpoint request headers when creating new request
This prevents headers intended for previous requests getting used
for new requests.

In particular this fixes the FAPI2baseline 'ensure holder of key
required' test, as this ensure a DPoP proof previously added to
the token endpoint (and potentially now invalid) is not reused when
attempting a call without a dpop proof.