Fix no such file ld-linux-x86-64.so.2 error in CI
The gcloud installation was failing with this error:

File "/builds/openid/conformance-suite/google-cloud-sdk/lib/third_party/google/protobuf/descriptor.py", line 48, in <module>
    from google.protobuf.pyext import _message
ImportError: Error loading shared library ld-linux-x86-64.so.2: No
such file or directory (needed by
/builds/openid/conformance-suite/google-cloud-sdk/lib/third_party/google/protobuf/pyext/_message.cpython-39-x86_64-linux-gnu.so)

This appears to have started with the latest release of gcloud;
explicitly ask for the older release instead.

I presume it's probably just a release that's not been correctly built
for alpine so probably we can move to the next release.

(I tried switching to the gcloud docker container instead, but that
failed later as a docker daemon needs to be running, and the
docker:git container seems to be the easiest way to achieve that.)