Common up id_token checks into a sequence
We have a lot of code that performs pretty identical checks; put it
all into one central condition sequence just to reduce the amount of
duplicate code, and perhaps also to make it clearer when a check is
deliberately different.

This overall isn't intended to add new checks, but a few of the places
the new sequence is added it contains more checks that the existing
code did. It also changes the ordering of checks in some places a
little, but again nothing appears to be a significant change.