Add validation of 'standard' id_token claims
For some reason we had a reasonable set of checks for the standard
OIDC claims if they were returned from userinfo, but not when they
were returned in the id_token.

Refactor such that we do the same sets of checks on all id_tokens
in all our different test profiles.

This is mostly inspired by the upcoming FAPI use cases where we
expect more identity claims to be returned.