FAPI2BaseOP: Add tests for claims parameter
This firstly adds a generic test for all profiles that runs if the OP
indicates that it supports at least one OpenID Connect identity claim
via the claims parameter (otherwise it goes to skipped).

The test requests the AS returns all the supported OpenID Connect
claims in both the id_token and userinfo, and expects the server to
return the requested claims in at least one of those locations. If it
doesn't, it fails with a warning (as there are lots of reasons the
claims might be not returned, not least that the user being used for
testing doesn't contain all the claims).

It includes a few random claims to make sure that unknown claims &
claims locations are correctly ignored by the server.

Then there is a specifc test added for IDMVP profile that requests 6
specific claims are returned in the id_token only, and fails if they
are not returned (as their standard is more specific).

In both cases, the claims are requested in a number of different ways.