Fix ValidateUserInfoStandardClaims to check address claims
I noticed a "Skipping unknown claim: address.country" output in the
log, and investigation showed that all the validators in the
ADDRESS_CLAIMS map had null entries, as the map was created
before the initialiser for VALIDATE_STRING etc.

Re-order the intialisers to fix this, and add a unit test that fails
without this fix.