FAPI2RPBrazil: Remove now unused parModules
This isn't referenced anymore, so remove it.

FAPI2BaselineID2ClientTestUnencryptedRequestObjectWithPAR is hence
also unused, so remove it.

In Brazil with FAPI2 request objects must always be unencrypted -
they were only required to be encrypted if sent via the front channel,
which as PAR is always used is FAPI2 is never the case.