FAPI2RP: Consolidate list of test modules
We currently have the test modules listed in two places; bring them
into sync to make the differences more obvious.

I'm not entirely sure the separate test plan for Brazil is still
necessary - it was originally added so there were both tests with and
without PAR, but 'with PAR' is now the only option.