FAPI2BaseRP: Remove some definitely not applicable tests
There is no longer a need for encrypted id tokens as there are no
front channel id tokens.

c_hash is only required in front channel id_tokens.

Missing/invalid nonces are not applicable as nonce is optional in the
code flow.

s_hash is not mentioned in FAPI2, so there's no requirement for
clients to cerify it.

The 'no at hash' test is just weird, it doesn't reference any standard
that says at hash is mandatory.

The 'no scope' test is unapplicable as it was only FAPI1 that required
the AS to *always* return the granted scopes; FAPI2 is back to RC6749
behaviour where the scope can be omitted if everything the client
requested was granted.

I suspect there are still some non-applicable tests; there are a number
that are intended to return invalid things in the authorization
endpoint id_token, and even if they returned the invalid id_token from
the token endpoint I'm not clear the client has to validate them.
(For example OIDCC 3.1.3.7-6 allows the client to skip verification of
the signature of a backchannel id_token.)