Remove EnsureUserInfoBirthDateValid
How that AbstractValidateOpenIdStandardClaims is checking birthdate
there's really no reason to have a standalone check for just
birthdate in the userinfo. Remove it and fold the unit tests into the
one for the standard claims.

Arguably this goes against our principle of having small focussed
conditions. I'm feeling kind of okay with this as it seems better than
the alternative (of adding a call to a 'check birthdate' condition to
every test where we validate an id_token).