FAPI2 OP: Remove options from baseline, created advanced test plan
As per
https://bitbucket.org/openid/fapi/issues/481/ecosystems-diverging-from-fapi2-baseline

FAPI2-baseline doesn't define any requirements around JARM / signed
request objects / etc so we can't really create certification tests
that use signed request objects for baseline.

We hence create a FAPI2-Advanced test plan (which contains all the
modules/options that were in the baseline test plan), and derive from
that a FAPI2-Baseline test plan that no longer has the JARM and signed
request object settings.

The likely outcome of this is we should also rename the test modules
to remove 'baseline' from the names, as the modules are used in
baseline and advanced, just with different variants.

Also add some handling of IDMVP to certificationProfileName.