FAPI2: Add first RP-against-OP CI test
More will need to be added, but one is a good start.

The sending of an ACR request in generic FAPI2 tests is removed -
it was causing a failure as the RP tests didn't return the acr, and
actually they're not required to as the requirement to support
acr has been removed from the FAPI standard at some point during
FAPI1 work.