FAPI2 OP against RP: Get signed request tests passing
Basically fix the RP test so that it passes the discovery test, as
per the addition I made to the CI in the previous commit.

This is mostly just adding missing values to the server metadata
the RP tests expose; I also took the opportunity to move some
generic values set by the FAPIBrazil...ServerConfiguration up
to being set in all cases, as they're required in all cases.
(Which in turn meant I needed to tweak the FAPI1 RP test in order
to preserve the existing behaviour there.)