Adding an ability to use alternate keystore implementations
The OBB project has a need to use private keys held in KMS for production tests. We have an implementation of KeyStore which
allows us to do this. The changes in this merge request allow us to configure such a keystore implementation, and make
calls to protected resources using this to load client keys for mtls authentication. The implementation itself does not need to
exist in this codebase, and there is scope for other similar implementations to be added fairly easily.

I have written tests to ensure that these changes do not impact how resources are called in the absence of any such alternate implementations

I have modified client key extraction slightly to allow for the configuration of said implementations, but have not yet extended this to
a second client cert. When we are happy that this mechanism works, that, along with similar hooks for client assertion signing etc. will
be forthcoming.