Tweak README + discourage PRs on github
Remove this 'this is the main repository' text from README.md as it
also appears on any forks and on the github mirror, which is kind of
confusing.

Create CONTRIBUTING and pull_request_template files for github to
discourage people from creating pull requests on the read-only mirror
we have on github and redirect them to gitlab.