FAPI RP: Generate server config with mtls_alias_endpoints in all cases
This makes it identical to the Brazil specific one and hence we remove
the Brazil specific one.

This means there's now an MTLS version of the userinfo endpoint
available (important for IDMVP).

Arguably we should have one test that doesn't have mtls_alias_endpoints
(i.e. lists the mtls versions, if they are needed, in the root of the
discovery document). However for now it's probably most important
to make sure the clients have noticed the existence of
mtls_alias_endpoints and are processing it.

That would need a little more care, as in the FAPI2+DPOP+private_key_jwt
'plain fapi' case, you would not use mtls at any endpoint - so the
generated config needs to depend on the variant.