Allow longer cert chains for incoming MTLS
We currently don't set SSLVerifyDepth in our apache configs, resulting
it in defaulting to 1.

At least one client in the IDMVP testing was supply a chain that
had 2 entries, resulting in apache rejecting it with an error:

Certificate Verification: Certificate Chain too long (chain has 2
certificates, but maximum allowed are only 1)

Although we have apache setup to not require any particular CA,
I presume it is still validating the chain presented is valid
and matches the cert.

Increase the depth, and (as it seems sensible) pass the chain down
to java in http headers, and make java include it in the logging
for the incoming http.

We continue not to do any certification of the chain/CA beyond what
Apache does, though perhaps we should look into this.