Ease configuration of admins
Stuart@Biza pointed out that the configuration is hard to override
from the command line as spring makes it a little hard to set a config
value to an empty string, resulting it in being harder to configure a
google admin group.

Just protect the different code paths with an explicit check as to
whether the admin issuer is set to gitlab or google.