FAPI2 RP: Add advanced test plan
As was recently done for OP, add an advanced test plan, and change
baseline to offer less optionality.

Note that the baseline client OP against RP can't currently be run
using run-tests-plan.py as the syntax for selecting particular
modules to run doesn't work if the modules have variants set in the
plan. Hence we run the advanced tests, but with the releveant options
selected to make it the same as baseline.