run-tests.sh: Include discovery in second IDMVP run too
I think this avoids any annoying case where the compare job sometimes
runs against the first IDMVP and something against the second.
(An alternative would be to add a specific config file for each
run, as that would make them different in the eyes of compare, but
if this way works it's probably simplest).