Make gitlab admin mapper more generic, allowing it to be used with Azure
Now any OIDC provider that can return a 'groups' array in the id_token
or userinfo can be used for the 'admin issuer'.

Reintroduces the 'googleIss' value just as that means the code can
more closely align with upstream. Still does not allow google to be
used for login here.