gitlab-ci.yml: Remove kubernetes:active
Doesn't currently seem to be true since we tried migrating to gitlab
agent - I've no idea what's going on, but with this present our jobs
don't run.

The reason this was originally added was to make the pipelines at
least look clean on forks that don't have a cluster setup. Maybe it
will start working again in the future and we can add it back.