gitlab-ci.yml: Add use_kube_context to enable gitlab agent
This brings us a bit into line with the upstream autodevops stuff:

https://gitlab.com/gitlab-org/gitlab/blob/master/lib/gitlab/ci/templates/Jobs/Deploy.gitlab-ci.yml

We're not currently using the gitlab agent (as it doesn't support
kubectl exec, which we rely on to insert the ci token into mongodb),
but this does at least make it possible to use it.

There's a lot of places in our CI where we're pulling in altered parts
of the autodevops code; we may want to try and gradually remove that,
and switch over our secrets to the official gitlab secrets mechanism
etc.

Also add commented out lines with the other upstream steps we're
currently missing.

relates to #1053