Increase production mongo ram, decrease review storage
Importing the production data into the new cluster was very slow with
1.5G of RAM, and fell over with an OOM error. Increasing to 4G made it
faster and resolved the OOM crash, and it's probably good to give
production a bit more ram anyway. Make this change for staging too, as
that also ends up with a large database.

We recently increased the production mongo storage size to 250G, and
change it to be an ssd. The review environments don't need that much
storage (and it's slightly pricey) so reduce their disk space.

This may cause a problem deploying to any existing review apps if
their branches are rebased (as you can't shrink an existing storage) -
if it does the fix is to stop the environment via the gitlab ui
(destroying it) then run a new pipeline for the branch to redeploy.