FAPI2: Fix resource endpoint mtls check
A simplification that had removed the complication of some OB use cases
having two resource endpoints resulted in the other cases not working.

Add a simpler path for the other cases, and fix a bug that was then
revealed in DisallowAccessTokenInQuery.

Also move the extraction so it only happens when we need it, rather
than in every test.