Enabling CORS for test plan paths
Addresses #1062, but is not an exact implementation of the behavior described
in https://gitlab.com/openid/conformance-suite/-/issues/1062#note_1017597450.

Instead, this implementation leverages built-in CORS features of Spring Boot,
which doesn't provide the granular control suggested in the comment due
to the way we use Spring Boot to handle HTTP requests and responses.