Enable CORS for all test plan paths except for a select few
Addresses #1062, but is not an exact implementation of the behavior
described in #1062 (comment 1017597450).

Instead, this implementation leverages built-in CORS features of Spring Boot,
which doesn't provide the granular control suggested in the comment due to the
way we use Spring Boot to handle HTTP requests and responses.

The change is not isolated to FAPI2 RP tests — they will apply to any test plan path,
except for paths with a trailing /authorize or /end_session_endpoint.
Please also note that paths that end with /check_session_iframe and /get_session_state
already had a configured CORS treatment which is left intact.