Fix some error responses including unescaped user content
Some error messages that looked like they should be a text/plain,
and are when using 'curl'. but when accessing them in a browser
because the Accept: header has text/html at the start Spring appears
to be decided to return it as text/html rather than text/plain,
but doesn't escape anything resulting in a possible XSS vulnerability.

Switch them to returning JSON instead, by returning a Map instead of
a plain string.

I looked at getting archunit to detect this pattern but ran into an
issue trying to get it to match constructors in a generic class which
I haven't yet found a solution for:

https://stackoverflow.com/questions/73372632/archunit-matching-on-a-constructor-for-a-generic-type