FAPIBrazil: Add opin option to Brazil specific test plans
We have some test plans that are only used in Brazil OpenBanking
currently, the OP DCR & the Brazil RP tests.

Rather than adding versions of these for Brazil OpenInsurance as well,
we remove the hardcoding of the profile and update
certificationProfileName() to prevent the plans being run for
non-Brazil profiles.