Feedback on download
Previously, the page would stay exactly the same after clicking the
submit button, which could be a little bit confusing if you didn't
notice that the zip download was happening.

This change provides a visual change of the dialog which provides
further instructions and serves as information regarding the downloaded file.