Require at least one finished and passed test
The certification package call will fail if it's invoked when there
are no applicable tests to include in it. The way it fails is a bit
unfortunate because it's a form post to an action that responds
with 400, leaving the end user with a blank screen.

This edit will light up the certification package button if there's
at least one FINISHED and PASSED test.