Fix 'Continue Plan' button not appearing for various test plans
The FAPI1 Brazil DCR OP & FAPI2-Advanced test plans were affected by
this.

The problem is that variants parts of the system treat the variants
being the empty object and being null as equivalent (as semantically
they are the way we use them). Unfortunately for these two test plans
(and likely some others too) the module list is returned with the
variant as an empty object and hence the code fails to find the
current test module & variant combo in the list, so can't figure out
which module is next, resulting in the button not appearing.