Delete test plan
This edit adds a button "Delete plan" to the plan detail page.
Clicking the button pops up a modal that explains that the action
cannot be undone and asks for a final button click confirmation,
upon which an async fetch call is done to the backend which deletes
the plan. Finally, the page is reloaded and a HTTP 404 error is
displayed as a confirmation to the user that the deletion has taken place.