Move the delete functionality and delete the tests as well
This commit moves the delete functionality from the LogApi to the
TestPlanApi where it belongs.

Additionally, deleting the test plan now also drills into the individual
tests and deletes them too, in order to not leave orphan tests around
and also because "deleting the plan" semantically implies to delete the
tests that it contains as well.

Finally, more information is added to the confirmation dialog regarding
what will be deleted and that it's forever gone.