Increase time between mongodb liveness checks
The current 10 seconds value seems like it may be resulting in
excessive CPU usage on the kubernetes cluster, particularly the review
apps one where:

mongosh --eval db.adminCommand('ping')

appeared to be taking ~2 seconds to complete and using relatively
significant amounts of cpu, it's quite often the top entry or two when
running 'top' on of the kubernetes nodes.

There are downsides with this, in that if the pod does die it'll take
longer for kubernetes to restart it - but as that seems to be an
infrequent event it may not be an issue. Possibly we could use a
different value for production.