Add unittest for LoadRequestIdTokenClaims + fix issues
Added a general happy path unit test.

Throw an error with claims.id_token.claims is not an object.

As a result of report from ConnectId, add a test where the contents
of the root claims element is a string instead of an object - fixing
this required a change to getElementFromObject() where it was ignoring
objects found with an incorrect type.