Upgrade mongo & mongochart
The current mongodb chart we're using has mysteriously vanished, as
have other ones that support the current mongodb version we're using
(4.4.x):

https://artifacthub.io/packages/helm/bitnami/mongodb/

which means that deploys currently fail.

This commit brings the helm chart onto the latest 5.0.x mongo (not the
latest stable release, but only one behind) and the latest
corresponding chart, as well as changing the docker compose files to
use the same version.

After the upgrade, any further deploys failed to happen, with the new
container stuck in 'creating' status; wading through the GCP console
logs eventually showed this error:

Multi-Attach error for volume
"pvc-dd232bb5-bffc-47bf-ac1e-7a958200f696" Volume is already used by
pod(s) review-miscfixes-cutowu-mongodb-6596cd6d57-9hmln

which led me to this issue:

https://github.com/helm/charts/issues/19716

which has the solution of changing the upgrade strategy from
'rollingupdate' to 'recreate' so that there's only ever one pod trying
to use the persistent volume with the database on it. I'm not clear
why we didn't need this previously; I can only presume it's an
undocumented change in the behaviour of the mongo helm chart.

A further issue arose on my local system:

Wrong mongod version","attr":{"error":"UPGRADE PROBLEM: Found an
invalid featureCompatibilityVersion document (ERROR: Location4926900:
Invalid featureCompatibilityVersion document in admin.system.version:
{ _id: \"featureCompatibilityVersion\", version: \"4.2\" }. See
https://docs.mongodb.com/master/release-notes/4.4-compatibility/#feature-compatibility.
Invalid value for featureCompatibilityVersiondocument in
admin.system.version, found 4.2, expected '4.4' or '4.9' or '5.0.'

I used this command to check if staging/production would have the same
issue:

db.adminCommand( { getParameter: 1, featureCompatibilityVersion: 1 } )

and both reported they are already on 4.4 (which as per above message
is compatible with the mongo 5.0 we're using).

We should probably look separately at updating that to 5.0 on
staging/production.

(instructions for getting shells on a local mongo shell are on the
build & run wiki page, and ones for getting a shell on a gcp mongo pod
are on the gcp maintenance page.)