Add checkstyle into maven test phase
This is being done primarily so errors are raised if wildcard
imports are used.

The .checkstyle.xml config file is based off the google one, but with
any rules that resulted in a substantial number of warnings from
our current code base removed.

A more considered approach to the configuration might be worth
looking at in the future.