Avoid checkstyle warning in AddPotentiallyBadBindingMessageToAuthorizationEndpointRequest
It doesn't like the unicode escapes.

I don't feel entirely confident in putting the emojis into the
text string here, as I suspect it'd make it hard to see it if an
editor decided to change the form etc. But checkstyle doesn't
warn if you add a comment to the line, so do that.