Set mongodb featureCompatibilityVersion on startup
As noted in
https://gitlab.com/openid/conformance-suite/-/merge_requests/1224
we're currently not very consistent with how we set this; I suspect
with my recent changes to upgrade to mongodb 5.0 then new review apps
are running with it set to 5.0, but existing review apps and staging +
production will still be running with it set to 4.4.

The only sensible way I can see to keep things consistent is to set
it on startup.

Now we're running mongodb 5.0 we should probably set it to 5.0, but
it might we wise to test a little more before we do that - so we're
setting it to 4.4 for now, which is a no-op apart from on new review
app environments.